+++ Steps/IO/Files/changelog.md
@@ -5,6 +5,7 @@
 - Upgrade all analyzers and remove deprecated analyzers.
 - Add the ExecutionTimings to DeleteFileOptions, MoveFileOptions, and WriteStreamToFileSystemOptions.
 - Add the ability to execute on the request or response side for DeleteFileStep, MoveFileStep, and WriteStreamToFileSystemStep.
+- Correct analyzer warnings.
 
 ## 2.0.0
 - Upgrade to .NET 5.0